pyproject.toml
[tool.poetry]
name = "op-tcg"
version = "0.2.3"
description = ""
authors = ["fteutsch <florianteutsch@yahoo.de>"]
license = "MIT"
readme = "README.md"
packages = [{include = "op_tcg"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-dotenv = "^1.0.0"
pydantic = "2.10.0"
click = "^8.1.7"
google-cloud-bigquery = "^3.20.1"
pandas = "^2.2.2"
pyarrow = "^15.0.2"
db-dtypes = "^1.2.0"
google-cloud-storage = "^2.16.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
pandera = "^0.19.3"
tqdm = "^4.66.4"
timer = "^0.3.0"


[tool.poetry.group.crawler.dependencies]
requests = "^2.31.0"
scrapy = "^2.11.1"
beautifulsoup4 = "^4.12.3"
crawl4ai = "^0.7.4"


[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry.group.frontend.dependencies]
cssutils = "^2.11.1"
plotly = "^6.0.0"
python-fasthtml = "^0.12.12"
google-cloud-firestore = "^2.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
#
#[build-system]
#requires = ["poetry-core", "setuptools"]
#build-backend = "setuptools.build_meta"

[tool.poetry.scripts]
optcg = 'op_tcg.cli_app:app'
build-frontend = "build_frontend:build_frontend"